test-requirements.txt
Django==1.7
flake8
django-nose==1.3
djangorestframework==2.4.4
coverage
yanc
django-json-field==0.5.7
